// src/ui/app/root/application/models/Examination/ExamScheduleVM.ts
export interface IExaminationExamScheduleVM {
    examScheduleId: string;
    fromDate: Date;
    toDate: Date;
    statusId: number;
    loggerId: string;
    examTypeId: string;
    examName: string;
    courseId: string;
    courseName: string;
    campusId: string;
    campusName: string;
    programDetailId: string;
    description: string;
    classId: string;
    className: string;
    sessionId: string;
    session: string;
    campusProgramId: string;
}